+++ tests/testthat/test-autoplot.summary.seroincidence.by.R
@@ -1,37 +1,53 @@
-test_that("results are consistent", {
+test_that(
+  desc = "results are consistent",
+  code = {
 
-   library(dplyr)
-   library(ggplot2)
 
-   xs_data <-
-     sees_pop_data_pk_100
+    library(dplyr)
+    library(ggplot2)
 
-   curve <-
-     typhoid_curves_nostrat_100 |>
-     filter(antigen_iso %in% c("HlyE_IgA", "HlyE_IgG"))
+    xs_data <-
+      sees_pop_data_pk_100
 
-   noise <-
-     example_noise_params_pk
+    curve <-
+      typhoid_curves_nostrat_100 |>
+      filter(antigen_iso %in% c("HlyE_IgA", "HlyE_IgG"))
 
-   est2 <- estimate_scr_by(
-     strata = c("catchment"),
-     pop_data = xs_data,
-     curve_params = curve,
-     noise_params = noise,
-     curve_strata_varnames = NULL,
-     noise_strata_varnames = NULL,
-     antigen_isos = c("HlyE_IgG", "HlyE_IgA"),
-     num_cores = 2 # Allow for parallel processing to decrease run time
-   )
+    noise <-
+      example_noise_params_pk
 
-   est2sum <- summary(est2)
+    est2 <- estimate_scr_by(
+      strata = c("catchment", "ageCat"),
+      pop_data = xs_data,
+      curve_params = curve,
+      noise_params = noise,
+      curve_strata_varnames = NULL,
+      noise_strata_varnames = NULL,
+      antigen_isos = c("HlyE_IgG", "HlyE_IgA"),
+      num_cores = 2 # Allow for parallel processing to decrease run time
+    )
 
-   plot1 <- autoplot(est2sum, "catchment", CI = TRUE)
+    est2sum <- summary(est2)
 
-   plot1 |> vdiffr::expect_doppelganger(title = "strat-est-plot-CI")
+    plot1 <- autoplot(est2sum,
+                      xvar = "ageCat",
+                      type = "scatter",
+                      dodge_width = 0.1,
+                      color_var = "catchment",
+                      CI = TRUE)
 
-   plot2 <- autoplot(est2sum, "catchment", CI = FALSE)
+    plot1 |> vdiffr::expect_doppelganger(title = "strat-est-plot-CI")
 
-   plot2 |> vdiffr::expect_doppelganger(title = "strat-est-plot-no-CI")
+    plot2 <- autoplot(est2sum,
+                      xvar = "ageCat",
+                      type = "scatter",
+                      CI = TRUE,
+                      dodge_width = 0.1,
+                      group_var = "catchment",
+                      color_var = "catchment")
 
-})
+    plot2 |> vdiffr::expect_doppelganger(title = "strat-est-plot-CI-lines")
+
+
+  }
+)