+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## New features
 
-* Added `autoplot.summary.estimate_scr_by()` to create horizontal bar charts of stratified results (#397)
+* Extended `autoplot.summary.seroincidence.by()` to include types for either scatter or bar plots of stratified results (#397)
+* added option to add lines using `group_var` input to `autoplot.summary.seroincidence.by()`
 * `autoplot.pop_data(type = "age-scatter")` now shows legend at bottom (#407)
 * `autoplot.pop_data(type = "age-scatter")` now facets by antigen isotype (#406)
 * Rename `est.incidence.by()` to `estimate_scr_by()` (#389)